[CHG] refactoring
